refactor: Split server and client auth
